mach-tegra: tegra_mpdecision: use own workqueues
This will alter mpdecision to alloc it's own wq rather than using
the predefined queue which is a shared resource.
Since we are using a lot of delays in here they could slow down
other users of the workqueue, which should be avoided.
Create 2 workqueues to avoid this, one for normal operations and
one for lpcpu checks while suspended. Both threads are not bound
to the issuer CPU (=> high-priority), have rescue worker tasks
and are freezable.
Also destroy the workqueues on exit, and rename tegra_mpdec to
tegra_mpdec_init.

Signed-off-by: Dennis Rassmann <showp1984@gmail.com>